+++ src/main/java/id/ac/ui/cs/advprog/functionality/model/Order.java
@@ -20,7 +20,7 @@ public class Order {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    private Long totalPrice;
+    private double totalPrice;
 
     private String shippingAddress;
